VCV-211: * Switched proxy.ts to a global import for cookie names * Switched login and signup pages to use user authentication cookie instead of useState to determine whether to display login/signup or send verification email; they no longer have 'use client' * Changed email verification 'email sent' message * Renamed email-verification-form to email-verification-prompt as it does not include a form
