+++ src/features/auth/components/signup-form.tsx
@@ -20,7 +20,6 @@ import {
     type NetworkError,
 } from '@/lib/network/network-error';
 import type { Result } from '@/lib/result/result';
-import { useRouter } from 'next/navigation';
 import { Eye, Lock, Mail } from 'lucide-react';
 import { useState } from 'react';
 import {
@@ -31,12 +30,9 @@ import {
     SelectValue,
 } from '@/components/ui/select';
 import { useGetPrograms } from '@/api/program/hooks/use-get-programs';
+import { useRouter } from 'next/navigation';
 
-interface SignupFormProps {
-    setShowVerify: (value: boolean) => void;
-}
-
-export default function SignupForm({ setShowVerify }: SignupFormProps) {
+export default function SignupForm() {
     const router = useRouter();
     const [showPassword, setShowPassword] = useState(false);
     const [showConfirmPassword, setShowConfirmPassword] = useState(false);
@@ -77,11 +73,6 @@ export default function SignupForm({ setShowVerify }: SignupFormProps) {
             console.error('Signup failed', signupResult);
             return;
         }
-
-        setShowVerify(true);
-        return;
-
-        router.replace('/');
         router.refresh();
     }
